Renamed two functions and new functions are added
The functions insert_last() and del_last() are renamed to append() and pop(). Also the function pop() returns the popped element at end.
Added functions count() and indexOf(). The count() function returns the number of occurance of an element.
The indexOf() function returns the positive index of first occurance of an element. If the element does not exist in list then shows ValueError and returns -1.